Feat: Refactor add page into drawer for adding tenant allow/block list entries
